requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
bcrypt==4.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
chromedriver-autoinstaller==0.6.4
click==8.1.7
distro==1.9.0
ecdsa==0.19.0
fastapi==0.115.2
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
jiter==0.6.1
openai==1.52.0
outcome==1.3.0.post0
packaging==24.1
passlib==1.7.4
pillow==11.0.0
psycopg2==2.9.10
pyasn1==0.6.1
pydantic==2.9.2
pydantic_core==2.23.4
PySocks==1.7.1
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.12
requests==2.32.3
rsa==4.9
selenium==4.25.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
starlette==0.40.0
tqdm==4.66.5
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.0
websocket-client==1.8.0
wsproto==1.2.0
